# Conflicts:
#	package/commerce/src/OroB2B/Bundle/AccountBundle/Resources/views/layouts/default/orob2b_account_frontend_account_user_role_update/layout.yml
#	package/commerce/src/OroB2B/Bundle/AlternativeCheckoutBundle/Resources/views/layouts/default/orob2b_checkout_frontend_checkout/ajax/payment.yml
#	package/commerce/src/OroB2B/Bundle/AlternativeCheckoutBundle/Resources/views/layouts/default/orob2b_checkout_frontend_checkout/page/payment.yml
#	package/commerce/src/OroB2B/Bundle/PaymentBundle/Layout/DataProvider/PaymentMethodsProvider.php
#	package/commerce/src/OroB2B/Bundle/ProductBundle/Resources/config/services.yml
#	package/commerce/src/OroB2B/Bundle/SEOBundle/Resources/views/layouts/default/orob2b_product_frontend_product_index/layout.yml
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Layout/DataProvider/FrontendShoppingListProductUnitsQuantityProvider.php
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Resources/config/services.yml
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Resources/views/layouts/default/orob2b_shopping_list_frontend_view/view.yml
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Tests/Unit/Layout/DataProvider/FrontendShoppingListProductUnitsQuantityProviderTest.php
#	package/platform/src/Oro/Bundle/UIBundle/Resources/views/layouts/base/page/layout.yml

+++ src/Oro/Bundle/WindowsBundle/Resources/public/js/widget/dialog-widget.js
@@ -26,7 +26,8 @@ define(function(require) {
             type: 'dialog',
             dialogOptions: null,
             stateEnabled: true,
-            incrementalPosition: true
+            incrementalPosition: true,
+            preventModelRemoval: false
         }),
 
         // Windows manager global variables
@@ -51,7 +52,6 @@ define(function(require) {
         },
 
         listen: {
-            'adoptedFormResetClick': 'remove',
             'widgetRender': 'onWidgetRender',
             'widgetReady': 'onContentUpdated',
             'page:request mediator': 'onPageChange'
@@ -140,7 +140,7 @@ define(function(require) {
                 return;
             }
             dialogManager.remove(this);
-            if (this.model) {
+            if (this.model && !this.options.preventModelRemoval) {
                 this.model.destroy({
                     error: _.bind(function(model, xhr) {
                         // Suppress error if it's 404 response and not debug mode
@@ -154,6 +154,8 @@ define(function(require) {
 
             // need to remove components in widget before DOM will be deleted
             this.disposePageComponents();
+            _.invoke(this.subviews, 'dispose');
+
             if (this.widget) {
                 this.widget.remove();
                 delete this.widget;
@@ -231,6 +233,10 @@ define(function(require) {
             }
         },
 
+        _onAdoptedFormResetClick: function() {
+            this.remove();
+        },
+
         /**
          * Removes dialog widget
          */
@@ -270,6 +276,7 @@ define(function(require) {
 
         _clearActionsContainer: function() {
             this.widget.dialog('actionsContainer').empty();
+            this.actionsEl = null;
         },
 
         _renderActions: function() {
@@ -311,12 +318,10 @@ define(function(require) {
             }, this));
         },
 
-        _afterLayoutInit: function() {
+        _renderHandler: function() {
             this.resetDialogPosition();
             this.widget.closest('.invisible').removeClass('invisible');
-            if (this.deferredRender) {
-                this._resolveDeferredRender();
-            }
+            this.trigger('widgetReady', this);
         },
 
         _initAdjustHeight: function(content) {
@@ -419,6 +424,7 @@ define(function(require) {
             this.internalSetDialogPosition(position, leftShift, topShift);
             this.leftAndWidthAdjustments(dialog, containerEl);
             this.topAndHeightAdjustments(dialog, containerEl);
+            this.widget.trigger('dialogreposition');
         },
 
         leftAndWidthAdjustments: function(dialog, containerEl) {
